refactor PackageDialog/State.tsx, split into atomic inputs states
