+++ catalog/app/containers/Bucket/PackageDialog/index.ts
@@ -9,7 +9,7 @@ export * from './MetaInput'
 export * from './PackageCreationForm'
 export * from './PackageDialog'
 export * from './Skeleton'
-export { useContext, Provider, isPackageHandle } from './state'
+export { useContext, Provider, isPackageHandle } from './State'
 export { default as SubmitSpinner } from './SubmitSpinner'
 export * from './Uploads'
 export { default as useInitialActions } from './actions'
